IS_2025/03_Generate_embedding/requirements.txt
pandas
pyarrow
fastparquet
qdrant_client
python-dotenv
sentence-transformers